+++ server/src/main/java/codestates/frogroup/indiego/global/security/auth/handler/MemberAccessDeniedHandler.java
@@ -1,6 +1,6 @@
 package codestates.frogroup.indiego.global.security.auth.handler;
 
-import codestates.frogroup.indiego.global.security.auth.utils.ErrorResponder;
+import codestates.frogroup.indiego.global.security.auth.utils.Responder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
@@ -17,6 +17,6 @@ public class MemberAccessDeniedHandler implements AccessDeniedHandler {
     @Override
     public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException {
         log.warn("Forbidden error happened: {}", accessDeniedException.getMessage());
-        ErrorResponder.sendErrorResponse(response, HttpStatus.FORBIDDEN);
+        Responder.sendErrorResponse(response, HttpStatus.FORBIDDEN);
     }
 }
