+++ server/src/main/java/codestates/frogroup/indiego/global/security/auth/handler/MemberAuthenticationEntryPoint.java
@@ -1,6 +1,6 @@
 package codestates.frogroup.indiego.global.security.auth.handler;
 
-import codestates.frogroup.indiego.global.security.auth.utils.ErrorResponder;
+import codestates.frogroup.indiego.global.security.auth.utils.Responder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.AuthenticationException;
@@ -18,7 +18,7 @@ public class MemberAuthenticationEntryPoint implements AuthenticationEntryPoint 
     public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
         Exception exception = (Exception) request.getAttribute("exception");
         logExceptionMessage(authException, exception);
-        ErrorResponder.sendErrorResponse(response, HttpStatus.UNAUTHORIZED);
+        Responder.sendErrorResponse(response, HttpStatus.UNAUTHORIZED);
     }
 
     private void logExceptionMessage(AuthenticationException authException, Exception exception){
