+++ server/src/main/java/codestates/frogroup/indiego/global/security/auth/utils/Responder.java
@@ -1,0 +1,41 @@
+package codestates.frogroup.indiego.global.security.auth.utils;
+
+import codestates.frogroup.indiego.domain.member.entity.dto.MemberDto;
+import codestates.frogroup.indiego.global.exception.BusinessLogicException;
+import codestates.frogroup.indiego.global.exception.ErrorResponse;
+import codestates.frogroup.indiego.global.exception.ExceptionCode;
+import codestates.frogroup.indiego.global.security.auth.userdetails.AuthMember;
+import com.google.gson.Gson;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class Responder {
+    public static void sendErrorResponse(HttpServletResponse response, HttpStatus status) throws IOException {
+        Gson gson = new Gson();
+        ErrorResponse errorResponse = ErrorResponse.of(status);
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setStatus(status.value());
+        response.getWriter().write(gson.toJson(errorResponse, ErrorResponse.class));
+    }
+
+    public static void sendErrorResponse(HttpServletResponse response, ExceptionCode code) {
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        throw new BusinessLogicException(code);
+    }
+
+    public static void loginSuccessResponse(HttpServletResponse response, AuthMember authMember) throws IOException {
+        Gson gson = new Gson();
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        MemberDto.LoginResponse loginResponse = MemberDto.LoginResponse.builder()
+                .id(authMember.getId())
+                .email(authMember.getEmail())
+                .role(authMember.getRoles().get(0))
+                .build();
+
+        response.getWriter().write(gson.toJson(loginResponse, MemberDto.LoginResponse.class));
+    }
+
+}
